requirements.txt
matplotlib
numpy
scipy
tensorboard
torch
torchvision
tqdm
rerun-sdk
wandb
